% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MLLR1}
\alias{MLLR1}
\alias{qmllr1}
\alias{pmllr1}
\alias{rmllr1}
\title{The MLLR_1 distribution}
\usage{
qmllr1(p, M, k, precision = 0.01, lower = TRUE, log_p = FALSE, force = FALSE)

pmllr1(q, M, k, precision = 0.01, lower = TRUE, log_p = FALSE, force = FALSE)

rmllr1(n, M, k)
}
\arguments{
\item{p}{vector of probabilities}

\item{M}{the first parameter for the MLLR_1 distributions}

\item{k}{the second parameter for the MLLR_1 distribution}

\item{precision}{The requested level of precision for the outputs of qmllr1 and pmllr1 functions, in terms of the estimated standard deviation of the output. For example precision of 0.01 will output values with the standard deviation of approximately equal to 0.01.}

\item{lower}{logical; if TRUE, probabilities are P\link{X <= x}, otherwise, P\link{X > x}.}

\item{log_p}{logical; if TRUE, probabilities p are given as log(p).}

\item{force}{logical; if TRUE, the function will run regardless of how long it will take. If FALSE, the function will ask if you want to continue, stop, or give a new precision value whenever the expected run time is longer than 15 seconds.}

\item{q}{vector of quantiles}

\item{n}{number of draws}
}
\value{
a list consisting of the numeric vector of quantiles and the precision (numeric) used.
}
\description{
Quantile function, distribution function, and random generation for the MLLR_1 distribution family. See Park (2023) for information about the MLLR distributions.
}
\examples{
qmllr1(.99, 5, 2)
qmllr1(c(.01, .05, .95, .99), 10, 2.3)
qmllr1(c(.01, .05, .95, .99), 10, 2.3, precision=0.01, lower=TRUE)
pmllr1(c(-8.3, -5.9), 8, 1)
pmllr1(c(-8.3, -5.9), 8 ,1, force=TRUE)
rmllr1(10, 7, 2)
}
