% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclle.R
\name{new_mclle}
\alias{new_mclle}
\title{Monte Carlo log likelihood estimator (MCLLE) class}
\usage{
new_mclle(llest, param = NULL)
}
\arguments{
\item{llest}{A numeric vector of log likelihood estimates}

\item{param}{A numeric vector of one-dimensional parameter values (optional).}
}
\value{
A class 'mclle' object

Constructor for a class 'mclle' object
}
\description{
Monte Carlo log likelihood estimator (MCLLE) class
}
