% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siblle.R
\name{new_siblle}
\alias{new_siblle}
\title{Constructor for a class 'siblle' object}
\usage{
new_siblle(llest, params = NULL)
}
\description{
Constructs a new class 'siblle' object. Note that the 'siblle' function uses the 'new_siblle' function to construct an object and use an interval validator function to check the correctness of the object specification.
}
