% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simll.R
\name{new_simll}
\alias{new_simll}
\title{Constructor for a class 'simll' object}
\usage{
new_simll(ll, params = NULL, weights = NULL)
}
\description{
Constructs a new class 'simll' object. Note that the 'simll' function uses the 'new_simll' function to construct an object and use an interval validator function to check the correctness of the object specification.
}
