% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qscl}
\alias{qscl}
\title{The SCL Distribution}
\usage{
qscl(p, M, precision, lower = TRUE, log = FALSE)
}
\arguments{
\item{p}{probability}

\item{M}{parameter to the SCL distribution}

\item{precision}{The requested level of precision for the outputs of qscl and pscl functions, as the number of decimal places. See Details for more information. (Details: precision can be any positive real, and checks if the standard error of the output is not greater than 10^{-precision}.)}
}
\value{
b b
}
\description{
Quantile function, distribution function, and random generation for the SCL distribution. See Park and Won (2022) for information about the SCL distribution.
}
\examples{
asdf5
}
