% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SCLdist}
\alias{SCLdist}
\alias{qscl}
\alias{pscl}
\alias{rscl}
\title{The SCL Distribution}
\usage{
qscl(p, M, k, precision = 0.01, lower = TRUE, log_p = FALSE, force = FALSE)

pscl(q, M, k, precision = 0.01, lower = TRUE, log_p = FALSE, force = FALSE)

rscl(n, M, k)
}
\arguments{
\item{p}{vector of probabilities}

\item{M}{the first parameter for the SCL distribution}

\item{k}{the second parameter for the SCL distribution}

\item{precision}{The requested level of precision for the outputs of qscl and pscl functions, in terms of the estimated standard deviation of the output. For example precision of 0.01 will output values with the standard deviation of approximately equal to 0.01.}

\item{lower}{logical; if TRUE, probabilities are P[X <= x], otherwise, P[X > x].}

\item{log_p}{logical; if TRUE, probabilities p are given as log(p).}

\item{force}{logical; if TRUE, the function will run regardless of how long it will take. If FALSE, the function will ask if you want to continue, stop, or give a new precision value whenever the expected run time is longer than 15 seconds.}

\item{q}{vector of quantiles}

\item{n}{number of draws}
}
\value{
numeric vector of quantiles.
}
\description{
Quantile function, distribution function, and random generation for the SCL distribution. See Park and Won (2022) for information about the SCL distribution.
}
\examples{
qscl(.99, 5, 2)
qscl(c(.01, .05, .95, .99), 10, 2.3)
qscl(c(.01, .05, .95, .99), 10, 2.3, precision=0.01, lower=TRUE)
pscl(c(-8.3, -5.9), 8, 1)
pscl(c(-8.3, -5.9), 8 ,1, force=TRUE)
rscl(10, 7, 2)
}
