% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siblle.R
\name{siblle}
\alias{siblle}
\title{Simulation based log likelihood estimator (SIBLLE) class}
\usage{
siblle(llest, params = NULL)
}
\arguments{
\item{llest}{A numeric vector of log likelihood estimates}

\item{params}{A numeric vector of one-dimensional parameter values (optional, used if the log likelihood estimates are obtained at more than one parameter values).}
}
\value{
A class 'siblle' object
}
\description{
Simulation based log likelihood estimator (SIBLLE) class
}
