% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simll.R
\name{simll}
\alias{simll}
\title{Simulation Log Likelihood class}
\usage{
simll(ll, params = NULL, weights = NULL)
}
\arguments{
\item{ll}{A matrix vector of simulation log likelihoods. The (m,i)-th entry is given by the log density of the observation y_i given the simulation X(theta_m).}

\item{params}{A numeric vector of one-dimensional parameter values (optional, can be omitted if simulation log likelihoods are obtained at only one parameter value.)}

\item{weights}{A numeric vector of weights, inversely proportional to the variance of simulation log likelihoods (optional)}
}
\value{
A class 'sll' object
}
\description{
Simulation Log Likelihood class
}
