% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_pomp.R
\name{SV_pomp}
\alias{SV_pomp}
\title{Stochastic volatility model}
\usage{
SV_pomp(t0, times, kappa, gamma, tau, sim_seed = NULL)
}
\arguments{
\item{t0:}{time zero where the rinit is called}

\item{times:}{observation times}

\item{kappa:}{a parameter in the model (between 0 and 1, close to 1 means the volatility has high autocorrelation)}

\item{tau:}{a parameter in the model (positive, the standard deviation of volatility at a given time point)}

\item{sim_seed:}{random seed for data generation}
}
\description{
\code{SV_pomp} defines a stochastic volatility model used as an example in the package vignette. The model is defined using the \code{R} package \code{pomp}.
}
\details{
The model is described as follows.

Observations: r_t = exp(s_t)*W_t where W_t~t_5 (t-distribution with 5 degrees of freedom)

Latent stochastic volatility: s_t = kappa\emph{s_t + tau}sqrt(1-kappa^2)\emph{V_t,  V_t~N(0,1),  for t>1
s_1 = tau}V_1,  V_1~N(0,1) for t=1.
}
