% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\alias{ci.mclle}
\title{Confidence intervals constructed using Monte Carlo log likelihood estimates in simulation-based inference}
\usage{
\method{ci}{mclle}(
  mclle,
  level,
  type = NULL,
  ci = NULL,
  param.at = NULL,
  weights = NULL,
  fraction = NULL,
  center = NULL
)
}
\arguments{
\item{mclle}{A class 'mclle' object, containing the Monte Carlo log likelihood estimates and (optional) the parameter values for which those estimates were obtained.}

\item{level}{The conservative confidence level. Either a numeric (vector of length one) or a list of numerics (for constructing multiple confidence intervals).}

\item{type}{A character string indicating what type of situation is considered. One of "point", "regression", or "LAN". See Details.}

\item{ci}{A character string indicating the quantity for which a confidence interval is to be constructed. One of "loglik", "MLE", "information", or "parameter". See Details.}

\item{param.at}{If 'ci' = "loglik", a confidence interval is constructed for the value of the log likelihood function evaluated at 'param.at' (for the cases 'type' = "regression" or "LAN".)}

\item{weights}{An optional argument for the (non-relative) weights of the Monte Carlo log likelihood estimates for regression. Either a numeric vector of length equal to the 'mclle' object, or a character string equal to "tricube".}

\item{fraction}{An optional argument indicating the fraction of points with nonzero weights for the case where 'weights' is specified as "tricube".}

\item{center}{An optional argument indicating the center of the local regression for the case where 'weights' is specified as "tricube".}
}
\value{
A list consisting of the followings are returned.
\itemize{
\item{Monte Carlo maximum likelihood estimate,}
\item{a data frame of the lower and upper bounds of the confidence intervals and the corresponding (conservative) confidence levels,}
\item{the precision (0.01 or 0.001) for the (conservative) confidence levels.}
}
}
\description{
`ci` constructs conservative confidence intervals using Monte Carlo log likelihood estimates. See Park and Won (2023) for more information.
}
\details{
This is a generic function, taking a class 'mclle' object as the first argument.
Confidence intervals are constructed under the assumption that the Monte Carlo likelihood estimator whose values are given in the 'mclle' object is (approximately) normally distributed.
When 'level' is a list, a conservative confidence interval is constructed for each level specified in the list.

The constructed confidence intervals are conservative, in that all points for which there exists a null distribution that is not rejected at the specified confidence level will be included in the constructed interval. This complication arises because a point in the confidence interval does not fully specify a distribution (instead only defines a subspace of the parameter space.) See Park and Won (2023) for more detailed explanation.

The 'type' argument should be one of "point", "regression", or "LAN".
The case 'type' = "point" means that the 'mclle' object contains Monte Carlo log likelihood estimates for a single, fixed parameter value.
The case 'type' = "regression" means that the 'mclle' object contains Monte Carlo log likelihood estimates evaluated at a range of parameter values, specified by the 'param' attribute of the 'mclle' object. A local quadratic regression for the estimated log likelihood values will be used for constructing confidence intervals, where the x-axis values are given by the 'param' values of the 'mclle' object.
The case 'type' = "LAN" means that inference on the model parameter will be carried out under the local asymptotic normality (Le Cam and Yang, 2000) condition.
If the 'mclle' object has 'param' attribute whose length is equal to the length of the object, then 'type' defaults to "LAN".
If the 'mclle' object does not have 'param' attribute, then 'type' defaults to "point".

When 'type' = "point", 'ci' can only be "loglik".
In this case a conservative confidence interval for the log likelihood given the observed data is constructed.
When 'type' = "point", 'test' = "loglik" is assumed by default.

When 'type' = "regression", 'ci' can be "loglik", "MLE", or "information".
If 'ci' = "loglik", confidence intervals are constructed for the value of the log likelihood function evaluated at 'param.at' given the observed data.
If 'ci' = "MLE", confidence intervals are constructed for the location of the maximum likelihood estimate given the observed data.
If 'ci' = "information", confidence intervals are constructed for the Fisher information, which is assumed to be equal to (-2) times the value of the quadratic coefficient of the regression polynomial describing the mean of the MCLLE.
When 'type' = "regression", 'ci' = "MLE" is assumed by default.

When 'type' = "LAN", 'ci' can be "loglik", "MLE", "information", or "parameter".
If 'ci' is "loglik", "MLE", or "information", the output is the same as in the case where 'type' = "regression".
If 'ci' is "parameter", confidence intervals for the value of the model parameter are constructed under the local asymptotic normality assumption.
When 'type' = "LAN", 'ci' = "parameter" is assumed by default.

When quadratic regression is carried out, the weights for the Monte Carlo likelihood estimates can be supplied.  The weights can either be given as an attribute 'weights' of the 'mclle' object, or as a function argument 'weights', with the latter being used when both are supplied. In either case, 'weights' should be a numeric vector of length equal to that of 'mclle'. If 'weights' is given as a function argument, it can be specified alternatively as a character string "tricube". In this case, the tricube weight (see Cleveland, 1979) is used, and the specified 'fraction' of the points will have nonzero weights. If weights are not supplied in either locations, all weights are taken to be equal to 1.
It is important to note that the weights should NOT be normalized. Multiplying all weights by the same constant changes the local regression results. Roughly speaking, the variance of the error in the Monte Carlo log likelihood estimate is assumed to be sigma^2/(the weight for the point). See Park & Won (2023) for more information.
}
\references{
Park, J. and Won, S. (2023). Simulation-based inference for partially observed, implicitly defined models

Cleveland, W. S. (1979). Robust locally weighted regression and smoothing scatterplots. Journal of the American statistical association, 74(368), 829-836.

Le Cam, L. and Yang, G. L. (2000). Asymptotics in statistics: some basic concepts. Springer-Verlag, New York.
}
