% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclle.R
\name{ht}
\alias{ht}
\title{Hypothesis tests using Monte Carlo log likelihood estimates}
\usage{
ht(x, ...)
}
\arguments{
\item{mclle}{A class 'mclle' object, containing the Monte Carlo log likelihood estimates and (optional) the parameter values for which those estimates were obtained.}

\item{null.value}{The null value for the hypothesis test. Either a numeric vector or a list of numeric vectors.}

\item{type}{A character string indicating what type of test will be carried out. One of "point", "regression", or "LAN". See Details.}

\item{test}{A character string indicating the quantity to be tested about. One of "loglik", "moments", "MLE", "information", or "parameter". See Details.}

\item{weights}{An optional argument for the weights of the Monte Carlo log likelihood estimates for regression. Either a numeric vector of length equal to the 'mclle' object, or a character string equal to "tricube".}

\item{fraction}{An optional argument indicating the fraction of points with nonzero weights for the case where 'weights' is specified as "tricube".}

\item{center}{An optional argument indicating the center of the local regression for the case where 'weights' is specified as "tricube".}
}
\description{
`ht` outputs the results of a hypothesis test based on the Monte Carlo log likelihood estimates (see Park and Won (2023) for details).
}
\details{
This is a generic function, taking a class 'mclle' object x as the first argument.
The hypothesis test is carried out under the assumption that the Monte Carlo likelihood estimator whose values are given in the 'mclle' object is (approximately) normally distributed.

When 'null.value' is a list, hypothesis tests are carried out for each component of the list.

The 'type' argument should be one of "point", "regression", or "LAN".
The case 'type' = "point" means that the 'mclle' object contains Monte Carlo log likelihood estimates for a single, fixed parameter value.
The case 'type' = "regression" means that the 'mclle' object contains Monte Carlo log likelihood estimates evaluated at a range of parameter values, specified by the 'param' attribute of the 'mclle' object. A local quadratic regression for the estimated log likelihood values will be used for the hypothesis test, where the x-axis values are given by the 'param' values of the 'mclle' object.
The case 'type' = "LAN" means that inference on the model parameter will be carried out under the local asymptotic normality (Le Cam and Yang, 2000) condition.
If the 'mclle' object has 'param' attribute whose length is equal to the length of the object, then 'type' defaults to "LAN".
If the 'mclle' object does not have 'param' attribute, then 'type' defaults to "point".

When 'type' = "point", 'test' can only be "loglik" or "moments".
In this case 'test' = "loglik" means the hypothesis test \eqn{H_0: l = null.value} versus \eqn{H_1: l != null.value} will be performed.
If 'test' = "moments", a test about the mean and the variance of the Monte Carlo log likelihood estimator is conducted.
The 'null.value' should be a numeric vector of length two (the first component being the mean and the second being the variance), or a list of numeric vectors of length two.
When 'type' = "point", 'test' = "loglik" is assumed by default, unless the 'test' argument is supplied.

When 'type' = "regression", 'test' can be "loglik", "moments", "MLE", or "information".
If 'test' = "loglik", the test is about the value of the log likelihood function evaluated at 'param.at'.
If 'test' = "moments", the test is about the quadruple \eqn{a, b, c, sigma^2} where \eqn{a, b, c} are coefficients of the polynomial describing the mean of the Monte Carlo likelihood estimator (i.e., \eqn{l(\theta) = a + b theta + c theta^2}) and \eqn{sigma^2} is the variance of the MCLLE.
If 'test' = "MLE", the test is about the location of the maximum likelihood estimate.
If 'test' = "information", the test is about the Fisher information, which is twice the vaale of \eqn{c}, the quadratic coefficient of the mean function.
When 'type' = "regression", 'test' = "MLE" is assumed by default.

When 'type' = "LAN", 'test' can be "loglik", "moments", "MLE", "information", or "parameter".
If 'test' is "loglik", "moments", "MLE", or "information", the output is the same as in the case where 'type' = "regression".
If 'test' is "parameter", a test about the value of the model parameter is conducted under the local asymptotic normality assumption.
When 'type' = "LAN", 'test' = "parameter" is assumed by default.

When quadratic regression is carried out, the weights for the Monte Carlo likelihood estimates can be supplied.  The weights can either be given as an attribute 'weights' of the 'mclle' object, or as a function argument 'weights'. In both cases, 'weights' should be a numeric vector of length equal to that of 'mclle'. If 'weights' is given as a function argument to 'ht', it can be specified alternatively as a character string "tricube". In this case, the tricube weight (see Cleveland, 1979) is used, with fraction \eqn{f} of points with nonzero weights specified by 'fraction'. If weights are not supplied in either locations, all Monte Carlo log likelihood estimates are used with equal weights.
}
\references{
Park, J. and Won, S. (2023) Simulation-based inference for partially observed, implicitly defined models

Cleveland, W. S. (1979). Robust locally weighted regression and smoothing scatterplots. Journal of the American statistical association, 74(368), 829-836.
}
