% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siblle.R
\name{new_siblle}
\alias{new_siblle}
\title{Simulation based log likelihood estimator (SIBLLE) class}
\usage{
new_siblle(llest, param = NULL)
}
\arguments{
\item{llest}{A numeric vector of log likelihood estimates}

\item{param}{A numeric vector of one-dimensional parameter values (optional).}
}
\value{
A class 'siblle' object

Constructor for a class 'siblle' object
}
\description{
Simulation based log likelihood estimator (SIBLLE) class
}
