% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbi-package.R
\docType{package}
\name{sbi-package}
\alias{sbi}
\alias{sbi-package}
\title{sbi: Simulation Based Inference}
\description{
This package offers inference methods for the cases where the probability model can be simulated but its density function cannot be evaluated. Inference is made by using random estimates of log likelihood obtained via simulation of the underlying process at specified parameter values. The package enables estimation of data dependent statistics, such as the maximum likelihood estimate or the value of the log likelihood function at a given point, under certain (e.g., normal) distributional assumptions on the log likelihood estimates. Furthermore, it carries out parameter inference under local asymptotic normality (LAN) assumption. Estimation and uncertainty quantification can be carried out via "ht" (hypothesis testing) and "ci" (confidence interval) functions.
}
\references{
Park, J. (2023) On simulation based inference for implicitly defined models (in preparation)
}
\author{
\strong{Maintainer}: Joonha Park \email{j.park@ku.edu} (\href{https://orcid.org/0000-0002-4493-7730}{ORCID})

}
\keyword{internal}
